[attributes] sync macro with govuk_frontend_jinja
This macro is derived from govuk-frontend and doesn't do anything unique
to our design system. So we can use this existing version that
already works with Jinja.

I previously rewrote this macro to avoid the intermediate attributesHtml
variable instead of using the namespace() workaround. This was because
of scoping issues in jinja preventing outer variables from being
modified in a loop. However, it's simpler to maintain if we use the
namespace() workaround as it keeps the code closely matching the
upstream version.

This also gets rid of the weak equality checks which were a bit dodgy.
Initially I wanted to avoid using `is` and other syntax unsupported by
nunjucks, but this had no real benefit, and diverging from the original
macro makes maintenence harder.